% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4.R
\name{as.data.frame,OpenCRAVATModuleSet-method}
\alias{as.data.frame,OpenCRAVATModuleSet-method}
\title{convert some of the fields of module metadata to a data.frame}
\usage{
\S4method{as.data.frame}{OpenCRAVATModuleSet}(x)
}
\description{
convert some of the fields of module metadata to a data.frame
}
