% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_sqlite.R
\name{barplot_gene_ontology}
\alias{barplot_gene_ontology}
\title{make a gene ontology barplot over all variants}
\usage{
barplot_gene_ontology(
  con,
  kpev = c("EXP", "TAS"),
  ont = c("MF", "BP"),
  ncat = 15
)
}
\arguments{
\item{con}{SQLite connection to openCRAVAT .sqlite}

\item{kpev}{character() evidence codes to retain, defaulting to EXP and TAS}

\item{ont}{character() ontology codes to retain, defaulting to MF and BP}
}
\value{
a ggplot instance
}
\description{
make a gene ontology barplot over all variants
}
