% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_sqlite.R
\name{barplot_sequence_ontology}
\alias{barplot_sequence_ontology}
\title{make a sequence ontology barplot over all variants}
\usage{
barplot_sequence_ontology(con)
}
\arguments{
\item{con}{SQLite connection to openCRAVAT .sqlite}
}
\value{
a ggplot instance
}
\description{
make a sequence ontology barplot over all variants
}
