% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_req.R
\name{browse_oc_result}
\alias{browse_oc_result}
\title{browse report from OpenCRAVAT request}
\usage{
browse_oc_result(x)
}
\arguments{
\item{x}{output of `run_oc_req`}
}
\description{
browse report from OpenCRAVAT request
}
\note{
need to check that run is complete
}
