% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ch19to38}
\alias{ch19to38}
\title{a chain object for transforming hg19 to GRCh38 coordinates}
\format{
rtracklayer chain structure
}
\usage{
ch19to38
}
\description{
a chain object for transforming hg19 to GRCh38 coordinates
}
\examples{
ch19to38
}
\keyword{datasets}
