% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{connect_local_annotator}
\alias{connect_local_annotator}
\title{connect to local annotator SQLite database}
\usage{
connect_local_annotator(name = "clinvar")
}
\arguments{
\item{name}{character(1) OpenCRAVAT annotator name}
}
\description{
connect to local annotator SQLite database
}
