% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crx_demo}
\alias{crx_demo}
\title{annotation of mutations in OpenCRAVAT's crx format}
\format{
data.frame
}
\usage{
crx_demo
}
\description{
annotation of mutations in OpenCRAVAT's crx format
}
\note{
500 records retained
}
\examples{
crx_demo[1:5,1:6]
}
\keyword{datasets}
