% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_oc_job_status}
\alias{get_oc_job_status}
\title{get job status}
\usage{
get_oc_job_status(baseurl = "http://0.0.0.0:8080/", job_id)
}
\arguments{
\item{baseurl}{character(1) server url}

\item{job_id}{character(1) id}
}
\description{
get job status
}
