% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_sqlite.R
\name{get_oc_sqlite_content}
\alias{get_oc_sqlite_content}
\title{get SQLite content for a table}
\usage{
get_oc_sqlite_content(con, tablename, as.data.frame = TRUE)
}
\description{
get SQLite content for a table
}
