% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_sqlite.R
\name{get_oc_tab}
\alias{get_oc_tab}
\title{get an optionally thinned table}
\usage{
get_oc_tab(
  con,
  tablename = "variant",
  thincols = c("All Mappings", "Samples", "Gene Targets"),
  use_header = TRUE
)
}
\arguments{
\item{con}{an RSQLite connection}

\item{tablename}{character(1)}

\item{thincols}{character() open-cravat columns}

\item{use_header}{logical(1)}
}
\description{
get an optionally thinned table
}
