% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_sqlite.R
\name{header_tx}
\alias{header_tx}
\title{transform internal _header tables to char vectors}
\usage{
header_tx(con, tblname)
}
\arguments{
\item{con}{DBI connection}

\item{tblname}{character(1) table name}
}
\value{
vector of character
}
\description{
transform internal _header tables to char vectors
}
