% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{import_local_crx}
\alias{import_local_crx}
\title{import the 'crx' file with variant annotation}
\usage{
import_local_crx(path = local_oc_path(), job_id)
}
\arguments{
\item{path}{character(1) defaults to `local_oc_path()`}

\item{job_id}{character(1) required job id}
}
\value{
a data.frame
}
\description{
import the 'crx' file with variant annotation
}
