% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{install_oc_module}
\alias{install_oc_module}
\title{install in the ocbenv}
\usage{
install_oc_module(modname)
}
\arguments{
\item{modname}{character(1) 'name' as defined in OpenCRAVAT installation protocol}
}
\description{
install in the ocbenv
}
