% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{key_modules}
\alias{key_modules}
\title{enumerate modules that are installed by GUI when an empty installation is encountered}
\usage{
key_modules()
}
\description{
enumerate modules that are installed by GUI when an empty installation is encountered
}
