% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_local_annotators}
\alias{list_local_annotators}
\title{use reticulate/basilisk to list locally installed OpenCRAVAT annotators}
\usage{
list_local_annotators()
}
\description{
use reticulate/basilisk to list locally installed OpenCRAVAT annotators
}
