% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{local_oc_job_resources}
\alias{local_oc_job_resources}
\title{list local job folder content}
\usage{
local_oc_job_resources(path = local_oc_path(), job_id, ...)
}
\arguments{
\item{path}{character(1) defaults to `local_oc_path()`}

\item{job_id}{character(1) a job id}

\item{\dots}{passed to `dir()`}
}
\description{
list local job folder content
}
