% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{local_oc_jobids}
\alias{local_oc_jobids}
\title{list local job id folder names}
\usage{
local_oc_jobids(path = local_oc_path())
}
\arguments{
\item{path}{character(1) defaults to `local_oc_path()`}
}
\description{
list local job id folder names
}
