% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{local_oc_path}
\alias{local_oc_path}
\title{use reticulate to determine where locally installed OpenCRAVAT modules reside}
\usage{
local_oc_path()
}
\description{
use reticulate to determine where locally installed OpenCRAVAT modules reside
}
