% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{local_oc_sqlite_con}
\alias{local_oc_sqlite_con}
\title{get connection to SQLite resource generated by a job}
\usage{
local_oc_sqlite_con(path = local_oc_path(), job_id)
}
\arguments{
\item{path}{character(1) defaults to `local_oc_path()`}

\item{job_id}{character(1) required job id}
}
\description{
get connection to SQLite resource generated by a job
}
