% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{local_oc_variant_tbl}
\alias{local_oc_variant_tbl}
\title{get variant tbl generated by a job}
\usage{
local_oc_variant_tbl(path = local_oc_path(), job_id)
}
\description{
get variant tbl generated by a job
}
