% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_postable.R
\name{make_oc_POSTable}
\alias{make_oc_POSTable}
\title{convert a GRanges to a data.frame that can be written out as a POSTable file for OpenCRAVAT}
\usage{
make_oc_POSTable(gr, chain = NULL, genome = NULL)
}
\arguments{
\item{gr}{GRanges instance}

\item{chain}{rtracklayer chain for liftOver}

\item{genome}{character(1) genome tag}
}
\description{
convert a GRanges to a data.frame that can be written out as a POSTable file for OpenCRAVAT
}
\examples{
tf = tempfile()
mtb = make_oc_POSTable(gr38)
write.table(mtb, file=tf, sep="\t", col.names=FALSE, row.names=FALSE, quote=FALSE)
read.delim(tf, sep="\t", nrow=5, h=FALSE)
}
