% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nonco_var}
\alias{nonco_var}
\title{nonco_var: example output of queryOC -- variant in non-coding region}
\format{
httr GET result
}
\source{
demo run
}
\usage{
nonco_var
}
\description{
nonco_var: example output of queryOC -- variant in non-coding region
}
\note{
The value component of 'cookies' was deleted.
}
\examples{
names(httr::content(nonco_var))
}
\keyword{datasets}
