% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invoke.R
\name{ocapp}
\alias{ocapp}
\title{simple app}
\usage{
ocapp(cravat_cmd = "cravat", sqlite_to_home = TRUE)
}
\arguments{
\item{cravat_cmd}{character(1) string to invoke cravat}

\item{sqlite_to_home}{logical(1) if TRUE will copy generated sqlite to home, with basename of input filename as prefix to .sqlite}
}
\description{
simple app
}
