% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4.R
\docType{class}
\name{OpenCRAVATModuleSet-class}
\alias{OpenCRAVATModuleSet-class}
\title{define set of R representations of OpenCRAVAT module metadata}
\description{
define set of R representations of OpenCRAVAT module metadata
}
