% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4.R
\name{populate_module_set}
\alias{populate_module_set}
\title{use reticulate to query OpenCRAVAT module store and populate an R object with metadata about available resources}
\usage{
populate_module_set()
}
\description{
use reticulate to query OpenCRAVAT module store and populate an R object with metadata about available resources
}
