% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4.R
\name{show,OpenCRAVATModule-method}
\alias{show,OpenCRAVATModule-method}
\title{display OpenCRAVAT module metadata in R}
\usage{
\S4method{show}{OpenCRAVATModule}(object)
}
\arguments{
\item{object}{instance of OpenCRAVATModule}
}
\description{
display OpenCRAVAT module metadata in R
}
