% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4.R
\name{show,OpenCRAVATModuleSet-method}
\alias{show,OpenCRAVATModuleSet-method}
\title{display set of R representations of OpenCRAVAT module metadata}
\usage{
\S4method{show}{OpenCRAVATModuleSet}(object)
}
\description{
display set of R representations of OpenCRAVAT module metadata
}
