% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SO_map}
\alias{SO_map}
\title{Sequence Ontology abbreviations and tags used with OpenCRAVAT}
\format{
data.frame
}
\usage{
SO_map
}
\description{
Sequence Ontology abbreviations and tags used with OpenCRAVAT
}
\examples{
head(SO_map)
}
\keyword{datasets}
