% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SO_onto}
\alias{SO_onto}
\title{Sequence Ontology ingested from OBO}
\format{
ontologyIndex instance
}
\usage{
SO_onto
}
\description{
Sequence Ontology ingested from OBO
}
\note{
Ingested from so.obo at https://github.com/The-Sequence-Ontology/SO-Ontologies/Ontology_Files
}
\examples{
head(names(SO_onto))
head(SO_onto$name)
}
\keyword{datasets}
