% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTnT.R
\name{TnTdemo}
\alias{TnTdemo}
\title{Illustration of variant vizualization with TnT}
\usage{
TnTdemo(
  mae = NULL,
  sitecode = "ACC",
  chr = "11",
  viewstart = 6e+07,
  viewend = 7e+07,
  verbose = TRUE
)
}
\arguments{
\item{mae}{instance of MultiAssayExperiment, defaults to NULL}

\item{sitecode}{character(1) TCGA code, defaults to "ACC"}

\item{chr}{character(1) chromosome for variant request, defaults to "11"}

\item{viewstart}{numeric(1) leftmost basepair to view, defaults to 6e7}

\item{viewend}{numeric(1) rightmost basepair to view, defaults to 7e7}

\item{verbose}{logical(1) passed to curatedTCGAData::curatedTCGAData}
}
\value{
TnT::TnTGenome
}
\description{
Illustration of variant vizualization with TnT
}
\note{
Mutation component is extracted from MultiAssayExperiment, and
rowRanges method used to obtain addresses of recorded somatic mutations.
Non-SNV mutations are dropped.
}
