% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{var_in_tx}
\alias{var_in_tx}
\title{var_in_tx: example output of queryOC -- variant in transcribed region}
\format{
httr GET result
}
\source{
demo run
}
\usage{
var_in_tx
}
\description{
var_in_tx: example output of queryOC -- variant in transcribed region
}
\note{
The value component of 'cookies' was deleted.
}
\examples{
var_in_tx
str(httr::content(var_in_tx))
}
\keyword{datasets}
