% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{local_oc_variant_tbl}
\alias{local_oc_variant_tbl}
\title{get variant tbl generated by a job}
\usage{
local_oc_variant_tbl(path = local_oc_path(), job_id)
}
\arguments{
\item{path}{character(1) defaults to `local_oc_path()`}

\item{job_id}{character(1) required job id}
}
\description{
get variant tbl generated by a job
}
