% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{queryOC}
\alias{queryOC}
\title{single-variant query}
\usage{
queryOC(
  uname,
  passwd,
  chr = "chr1",
  pos = "12777320",
  ref = "G",
  alt = "T",
  annotators = c("clinvar", "dbsnp", "exac_gene", "go", "rvis")
)
}
\arguments{
\item{uname}{user name registered at run.opencravat.org; will use value of environment variable OCRAVAT_USERNAME if missing}

\item{passwd}{password for opencravat account; will use value of environment variable OCRAVAT_PASSWORD if missing}

\item{chr}{character(1) chromosome in chrn format}

\item{pos}{character(1) position, in hg38 coordinates}

\item{ref}{character(1) reference base}

\item{alt}{character(1) alternate base}

\item{annotators}{character() vector of annotator tags}
}
\description{
single-variant query
}
\examples{
queryOC # don't want to use auth
}
